backend/requirements.txt

fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.36
alembic==1.13.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
